Minor fixes from PR comments as well as `OSTimeoutHandler` class * `OSTimeoutHandler` class extended by a class to create Timeout for that instance of the class   * Start timeout with `startTimeout` and stop with `destroyTimeout`   * Extended by `NotificationGenerationJob`, which is extended by `AppNotificationGenerationJob` and `ExtNotificationGenerationJob` * Added a `isComplete` `boolean` to the `NotificationGenerationJob` class to make sure complete methods can not be called twice   * Use case would be that work inside of a handler takes longer than timeout, but also calls complete at the end   * `isComplete` would be set `true` already so no way to complete more than once * `AppNotificationGenerationJob` and `ExtNotificationGenerationJob` now have `synchronized` on complete methods
